checksum work in progress
added multiplex to mash a sequence and its checksum together before encryption
added de-multiplex to undo multiplex
added checksum-seq to add checksums into a sequence. it has a bug when the interval is one less than a multiple of the length of the seq.

added tests for these. one currently fails.