add tests for main and decrypt.
added some toplevel tests for main and doscovered that padding is leaking
into the decrypted files. This is a problem because fixing the leak will
complicate using this code for streams/pipes. The large block size used
is complicating things because of its padding requirements.

This needs to be fixed by replacing arbitrily large blocks that also make
up the on-disc/on-wire format with a seperated "presentation layer". more
to come on this...